docs PR-buildbot script: minor improvements
 - remember timestamp when docs build was started (instead of when it
   was finished), to avoid missing to issue a TODO for pushes during the
   build
 - check if index.html for the output really is in place before marking
   a PR build as done

+++ misc/scripts/docs_buildbot/update-docs-pr.sh
@@ -5,13 +5,20 @@ for FILE in `ls pull_request_docs/*.todo 2> /dev/null`
 do
     PR=${FILE##*/}
     PR=${PR%.*}
+    DATETIME=`date --iso-8601=seconds`
     bash $HOME/update-docs.sh -p $PR
     STATUS=`echo $?`
     echo $STATUS
     if [ ! "$STATUS" = "-1" ]
     then
-        touch $HOME/pull_request_docs/${PR}.done
-        rm $HOME/pull_request_docs/${PR}.todo
+        # double check that output is really where it should be before marking
+        # build as done (in case the docs build gets terminated from outside
+        # leading to a non-"-1" return code)
+        if [ -f $HOME/pull_request_docs/${PR}/index.html ]
+        then
+            touch -d "$DATETIME" $HOME/pull_request_docs/${PR}.done
+            rm $HOME/pull_request_docs/${PR}.todo
+        fi
     fi
     cp $HOME/update-docs-pr/log.txt $HOME/pull_request_docs/${PR}.log
     cp $HOME/update-docs-pr/log.txt $HOME/htdocs/docs/pull_requests/${PR}.log
